mem: Change ABTU_malloc to solve a cache line sharing issue.
Most objects in Argobots are allocated by malloc, so they often
happen to be in the same cache line and significantly decrease
performance by contentions. To avoid this issue, all objects in
Argobots are aligned by cache line size, and those sizes are of
multiple of cache line size so that any objects are allocated in
the same cache line.