pool: Implement CC-Queue for a fifo pool.
CC-Queue is used instead of a queue based on spinlock. CC-Queue performs
as good as a spinlock-based queue when contended threads do not exist,
while it performs better when queues are contended.

See the following paper for details.
P. Fatourou and N. D. Kallimanis. Revisiting the Combining Synchronization Technique, PPoPP '12.

This implementation is based on the following:
https://github.com/chaoran/fast-wait-free-queue