thread: move ABTI_local_set_thread and ABTI_local_set_task.
TLS variables ABTI_lp_local->p_thread and p_task must be updated whenever the
context is changed to a ULT. The previous rule is as follows:
- thread to thread: the caller must update.
- thread to sched : no update.
- sched to thread : the caller must update.
- sched to sched  : if scheduler is invoked, the caller must updates it.
                    Otherwise, no update.

This rule is very complicated and difficult to follow. This commit moves most
ABTI_local_set_thread and _task functions into the context switching functions
so that developers do not take care of these values.

Specifically, the rule is changed as follows:
- xxx-to-thread: the caller needs to update.
- xxx-to-sched : no update.

Manual updates are required in the following places because they are not invoked
by user-level context switches:
- Main thread ULT
- Tasklet

The following changes are nontrivial.

1. TLS value management on sched-sched context switch

The new rule changed the management policy of these TLS values on sched-sched
context switch (which happens only when the stackable scheduler feature is
enabled). These values were used only to know the previous ULT after context
switches in ABTI_xstream_schedule_thread. For example, in the sched-thread case,
from the viewpoint of the scheduler, the previous ULT which resumed this
scheduler may not be the same as one which this scheduler scheduled (which
happens only by scheduler bypassing).

However, according to the stackable scheduler concept, schedulers are always
linearly stacked, so the previous ULT may not be different from the ULT to which
the parent scheduler invoked previously. This commit uses this knowledge to get
the previous ULT on sched-sched context switch.

Note that, originally, p_thread and p_task are not properly managed for a
scheduler-type ULT and thus users may not call any Argobots functions using
ABTI_lp_local internally in a custom scheduler; they did not work even in the
previous Argobots when the stackable scheduler feature is turned off. After this
commit, they do not work even when the stackable scheduler feature is enabled.

2. Removing ABTI_local_set_thread in ABT_thread_join

Now all the context switch functions properly update ABTI_lp_local->p_thread if
the target ULT is of non-scheduler type, so the special handling in
ABT_thread_join becomes unnecessary.

Note that this control comes from the scheduler-bypass path in
ABTDI_thread_terminate, which missed ABTI_local_set_thread.