thread: invalidate unused contexts
Contexts of ULTs associated with OS-level threads (the main ULT and the main
scheduler ULTs) are not used, since they are run on OS-level threads.
Previously, such contexts are uninitialized, so any values can be assigned,
which makes debugging difficult.

This commit explicitly initializes these contexts by NULL. Note there is no
performance penalty in the critical paths because these contexts are created
only on initialization.