atomic: implement scalable LIFO structure
ABTI_sync_lifo is a scalable LIFO implementation that does not have the ABA
problem.  If an atomic tagged pointer operation is supported (i.e., most x86/64
with ICC, GCC, Clang, 64-bit ARM with GCC, and POWER 8 and 9 with XLC and GCC),
push and pop are lock-free.  If not, it falls back on a spinlock-based blocking
implementation.