unit: control a thread stack by p_parent of ABTI_unit
Previously a thread stack is represented as a stack of schedulers.  To control
the thread stack without a scheduler, this patch adds a new member variable
p_parent to ABTI_thread.  This is necessary to let a non-scheduler ULT schedule
other ULTs and tasks.

Note that this commit just adds an overhead.  The next commit removes the
counterpart in ABTI_sched.